correctly split the test_task.py file into integration / unit tests
