+++ tests/sentry/workflow_engine/test_task_integration.py
@@ -5,7 +5,6 @@ from sentry.issues.ingest import hash_fingerprint
 from sentry.issues.occurrence_consumer import _process_message
 from sentry.issues.status_change_consumer import update_status
 from sentry.issues.status_change_message import StatusChangeMessageData
-from sentry.models.activity import Activity
 from sentry.models.group import GroupStatus
 from sentry.models.grouphash import GroupHash
 from sentry.testutils.cases import TestCase
@@ -94,31 +93,3 @@ class IssuePlatformIntegrationTests(TestCase):
                 "workflow_engine.process_workflow.activity_update",
                 tags={"activity_type": ActivityType.SET_RESOLVED.value},
             )
-
-
-class WorkflowStatusUpdateHandlerTests(TestCase):
-    def test__no_detector_id(self):
-        """
-        Test that the workflow_status_update_handler does not crash
-        when no detector_id is provided in the status change message.
-        """
-        group = self.create_group(project=self.project)
-        activity = Activity(
-            project=self.project,
-            group=group,
-            type=ActivityType.SET_RESOLVED.value,
-            data={"fingerprint": ["test_fingerprint"]},
-        )
-
-        message = StatusChangeMessageData(
-            id="test_message_id",
-            project_id=self.project.id,
-            new_status=GroupStatus.RESOLVED,
-            new_substatus=None,
-            fingerprint=["test_fingerprint"],
-            detector_id=None,  # No detector_id provided
-        )
-
-        with mock.patch("sentry.workflow_engine.tasks.metrics.incr") as mock_incr:
-            workflow_status_update_handler(group, message, activity)
-            mock_incr.assert_called_with("workflow_engine.error.tasks.no_detector_id")
